+++ packages/ibm-products-community/src/components/Example/Example.stories.js
@@ -1,0 +1,12 @@
+import { getStoryTitle } from '../../../../ibm-products/src/global/js/utils/story-helper';
+import { Example } from '.';
+
+export default {
+  title: getStoryTitle(Example.displayName),
+  component: Example,
+  tags: ['autodocs'],
+};
+
+export const Default = {
+  args: {},
+};