+++ packages/ibm-products/src/components/Datagrid/useSelectRows.js
@@ -1,11 +1,10 @@
-/*
- * Licensed Materials - Property of IBM
- * 5724-Q36
- * (c) Copyright IBM Corp. 2020
- * US Government Users Restricted Rights - Use, duplication or disclosure
- * restricted by GSA ADP Schedule Contract with IBM Corp.
+/**
+ * Copyright IBM Corp. 2020, 2023
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
  */
-// @flow
+
 import React, { useLayoutEffect, useState } from 'react';
 import cx from 'classnames';
 import { TableSelectRow } from '@carbon/react';
@@ -30,14 +29,29 @@ const useSelectRows = (hooks) => {
       withSelectRows: true,
     });
   });
-  hooks.visibleColumns.push((columns) => [
-    {
-      id: selectionColumnId,
-      Header: (gridState) => <SelectAll {...gridState} />,
-      Cell: (gridState) => <SelectRow {...gridState} />,
-    },
-    ...columns,
-  ]);
+  hooks.visibleColumns.push((columns) => {
+    // Ensures that the first column is the row expander in the
+    // case of selected rows and expandable rows being used together
+    const newColOrder = [...columns];
+    const expanderColumnIndex = newColOrder.findIndex(
+      (col) => col.id === 'expander'
+    );
+    const expanderCol =
+      expanderColumnIndex > -1
+        ? newColOrder.splice(expanderColumnIndex, 1)
+        : [];
+    return [
+      ...(expanderColumnIndex > -1 && expanderCol && expanderCol.length
+        ? expanderCol
+        : []),
+      {
+        id: selectionColumnId,
+        Header: (gridState) => <SelectAll {...gridState} />,
+        Cell: (gridState) => <SelectRow {...gridState} />,
+      },
+      ...newColOrder,
+    ];
+  });
 };
 
 const useHighlightSelection = (hooks) => {
@@ -72,6 +86,7 @@ const SelectRow = (datagridState) => {
     columns,
     withStickyColumn,
     dispatch,
+    getRowId,
   } = datagridState;
 
   const [windowSize, setWindowSize] = useState(window.innerWidth);
@@ -93,7 +108,12 @@ const SelectRow = (datagridState) => {
     }
     onChange(event);
     onRowSelect?.(row, event);
-    handleToggleRowSelected(dispatch, row, event.target.checked);
+    handleToggleRowSelected({
+      dispatch,
+      rowData: row,
+      isChecked: event.target.checked,
+      getRowId,
+    });
   };
 
   const selectDisabled = isFetching || row.getRowProps().disabled;