+++ packages/ibm-products-styles/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.19.1](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.19.0...@carbon/ibm-products-styles@2.19.1) (2023-12-26)
+
+**Note:** Version bump only for package @carbon/ibm-products-styles
+
+
+
+
+
+# [2.19.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.18.1...@carbon/ibm-products-styles@2.19.0) (2023-12-19)
+
+
+### Bug Fixes
+
+* **Guidebanner:** allow adding 1rem space on the left ([#3965](https://github.com/carbon-design-system/ibm-products/issues/3965)) ([fcd0388](https://github.com/carbon-design-system/ibm-products/commit/fcd038879e19b6498c63d2848fc5ae25c74c086c))
+* **InlineTip:** allow adding 1rem space on the left, v2 ([#3969](https://github.com/carbon-design-system/ibm-products/issues/3969)) ([feaa1ce](https://github.com/carbon-design-system/ibm-products/commit/feaa1cecf0386213170ace8489cbfb0ebe45f500))
+
+
+### Features
+
+* Cards ai ([#3953](https://github.com/carbon-design-system/ibm-products/issues/3953)) ([e7ab94e](https://github.com/carbon-design-system/ibm-products/commit/e7ab94eca8dcb86ce7227f83c6ed0923c413df32))
+* side panel for create and edit + type check ([#3954](https://github.com/carbon-design-system/ibm-products/issues/3954)) ([8589f2e](https://github.com/carbon-design-system/ibm-products/commit/8589f2ef537d2001f1a04cb98b8dddb915f20ef8))
+
+
+
+
+
+## [2.18.1](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.18.0...@carbon/ibm-products-styles@2.18.1) (2023-12-12)
+
+
+### Bug Fixes
+
+* **Checklist:** add carbon scss import files ([#3951](https://github.com/carbon-design-system/ibm-products/issues/3951)) ([4be8bc2](https://github.com/carbon-design-system/ibm-products/commit/4be8bc2aa03793036769bb22a00d2b39abcece1a))
+
+
+
+
+
+# [2.18.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.17.2...@carbon/ibm-products-styles@2.18.0) (2023-12-07)
+
+
+### Bug Fixes
+
+* fixes 3890 ([#3891](https://github.com/carbon-design-system/ibm-products/issues/3891)) ([3b7b6fd](https://github.com/carbon-design-system/ibm-products/commit/3b7b6fd11223d2cc517cd59438f7449b33ea7375))
+* **WebTerminal:** Apply inline theming web terminal. ([#3888](https://github.com/carbon-design-system/ibm-products/issues/3888)) ([80a54db](https://github.com/carbon-design-system/ibm-products/commit/80a54dbf22bdaefe2b7965902268941dc702f624))
+
+
+### Features
+
+* **FilterSummary:** enable filter summary clear button to be placed inline ([#3906](https://github.com/carbon-design-system/ibm-products/issues/3906)) ([e154f89](https://github.com/carbon-design-system/ibm-products/commit/e154f89d4295f646522d7f0d20c49323c69b4c75))
+* new NonLinearReading component ([#3889](https://github.com/carbon-design-system/ibm-products/issues/3889)) ([962fbb6](https://github.com/carbon-design-system/ibm-products/commit/962fbb61ea2897db2b954afb8ebd5e721afe30c9))
+* port Guidebanner component w Carousel to v2 ([#3905](https://github.com/carbon-design-system/ibm-products/issues/3905)) ([e89a3f2](https://github.com/carbon-design-system/ibm-products/commit/e89a3f2575387e5f46338955c4aa4d07de663e3b))
+
+
+
+
+
+## [2.17.2](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.17.1...@carbon/ibm-products-styles@2.17.2) (2023-12-05)
+
+
+### Bug Fixes
+
+* **Datagrid:** expandable/nested review fixes, custom hook for keeping expander focus ([#3861](https://github.com/carbon-design-system/ibm-products/issues/3861)) ([ae33c58](https://github.com/carbon-design-system/ibm-products/commit/ae33c58fe13f5b352d782a6ba64e8d428c65c498))
+* published side panel error ([#3900](https://github.com/carbon-design-system/ibm-products/issues/3900)) ([b1487f6](https://github.com/carbon-design-system/ibm-products/commit/b1487f600e4d6be683ec16de3783a721dad38f2d))
+* remove file added in error ([#3882](https://github.com/carbon-design-system/ibm-products/issues/3882)) ([c596ea2](https://github.com/carbon-design-system/ibm-products/commit/c596ea2f76d702f7dcd7c3b88d2dcb582ab0499d))
+* update Carbon 11 compatible versions to latest ([#3880](https://github.com/carbon-design-system/ibm-products/issues/3880)) ([b96b1f2](https://github.com/carbon-design-system/ibm-products/commit/b96b1f2982336f15f504cf75d2d50f30c88c93a9))
+
+
+
+
+
+## [2.17.1](https://github.com/carbon-design-system/ibm-products/compare/@carbon/ibm-products-styles@2.16.0...@carbon/ibm-products-styles@2.17.1) (2023-11-30)
+
+
+### Bug Fixes
+
+* **CreateTearsheet:** fix step height to fill available space ([#3860](https://github.com/carbon-design-system/ibm-products/issues/3860)) ([3f0e0d9](https://github.com/carbon-design-system/ibm-products/commit/3f0e0d95cc57a8003f243e1469726c333809a285))
+* manually bump styles package version ([#3858](https://github.com/carbon-design-system/ibm-products/issues/3858)) ([a18b72a](https://github.com/carbon-design-system/ibm-products/commit/a18b72aa07a76db1429743c9d830b1d20acf7056))
+* **WebTerminal:** removed web terminal styling ([#3848](https://github.com/carbon-design-system/ibm-products/issues/3848)) ([e6fa105](https://github.com/carbon-design-system/ibm-products/commit/e6fa1050a6864f709df280638aa1875362dcc53a))
+
+
+
+
+
 # 2.16.0 (2023-11-28)